Privacy screen image file name can now be overridden on iOS with a preference in config.xml
If "privacyScreenImage" preference is set, it will be explicitly used as a privacy screen image. Virtually this could be an image set under Images.xcassets.
If "privacyScreenImage" preference is not set, plugin will try to guess the image file name just as before.

NB. For example when using story board launch images as launch images, Privacy Screen Plugin does not find the correct images and e.g. camera consent dialog result in a application freeze. If, however, "LaunchStoryboard" is defined as "privacyScreenImage" preference, this will override plugins guesswork and run smoothly.